+++ tests/test_negative.py
@@ -43,7 +43,7 @@ class FakeClient:
         def __init__(self, outer):
             self._outer = outer
 
-        def list(self, all=False):
+        def list(self, all=False, ignore_removed=False):
             return self._outer._containers
 
     @property